requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
facenet-pytorch>=2.5.0
mtcnn>=0.1.1
faiss-cpu>=1.7.0